WIP update Python & Django & stuffs

+++ apps/table/helper.py
@@ -1,12 +1,13 @@
 from apps.table.models import ROWS, COLS
 
+
 def snakish_to_coord(pos):
 	if pos >= ROWS * COLS:
 		raise RuntimeError("Woops. Too much Snake!")
 
-	y,x = 0,0
-	y = pos / COLS
-	if (y % 2 == 1):
+	y, x = 0, 0
+	y = int(pos / COLS)
+	if y % 2 == 1:
 		x = (COLS - 1) - (pos - (y * COLS))
 	else:
 		x = pos - (y * COLS)
